Edit + Delete

+++ src/command/DeleteItemCommand.java
@@ -6,16 +6,22 @@ import view.ItemManagementView;
 public class DeleteItemCommand implements Command {
     private ItemDAO itemDAO;
     private ItemManagementView view;
+    private String itemCode;
 
     public DeleteItemCommand(ItemDAO itemDAO, ItemManagementView view) {
         this.itemDAO = itemDAO;
         this.view = view;
     }
 
+    public void setItemCode(String itemCode) {
+        this.itemCode = itemCode;
+    }
+
     @Override
     public void execute() {
-        String itemCode = view.getItemCodeToDelete();
-        itemDAO.deleteItemByCode(itemCode);
-        view.showDeleteItemSuccess();
+        if (itemCode != null) {
+            itemDAO.deleteItemByCode(itemCode);
+            view.showDeleteItemSuccess();
+        }
     }
 }
